:bug: fix cyclic dependency and add tests
