+++ src/actions/coffee-roulette/try-again-coffee.ts
@@ -6,12 +6,12 @@ import { coffeeRoulette, CoffeeRouletteProps } from "./coffee-roulette"
 export const tryAgainCoffee = async (platform: Platform, data: ButtonActionProps): Promise<void> => {
   const i18n = await I18n.getInstance()
 
-  platform.updateMessage(data.responseUrl, i18n.translate("coffeeRoulette.tryAgain"))
+  await platform.updateMessage(data.responseUrl, i18n.translate("coffeeRoulette.tryAgain"))
 
   const coffeeProps: CoffeeRouletteProps = {
     communityId: data.communityId,
     userId: data.userId.id,
   }
 
-  coffeeRoulette(platform, coffeeProps)
+  return coffeeRoulette(platform, coffeeProps)
 }
