+++ src/models/platform/slack/views/views.ts
@@ -1,8 +1,8 @@
 import { InteractiveView, View } from "../../message"
-import { CoffeeRouletteMessageView } from "./coffee-roulette-message-view"
-import { CoffeeRouletteTryAgainView } from "./coffee-roulette-try-again"
 import { GratitudeMessageInteractiveView } from "./view-gratitude-message"
 import { GratitudeSummaryView } from "./view-gratitude-summary"
+import { CoffeeRouletteProps } from "../../../../actions/coffee-roulette/coffee-roulette"
+import { I18n } from "../../../../services/i18n/i18n"
 
 export { GratitudeSummaryViewProps } from "./view-gratitude-summary"
 // TODO: maybe add tests
@@ -29,8 +29,87 @@ export class SlackInteractiveBlock extends InteractiveView implements SlackBlock
       super()
     }
 
-  static coffeeRouletteMessage = CoffeeRouletteMessageView
-  static tryAgainCoffeeMessage = CoffeeRouletteTryAgainView
+  static coffeeRouletteMessage = async (data: CoffeeRouletteProps): Promise<SlackInteractiveBlock> => {
+    const i18n = await I18n.getInstance()
+    const options = { sender: `<@${data.userId}>`, text: data.text }
+    const blocks = [
+      {
+        "type": "section",
+        "text": {
+          "type": "mrkdwn",
+          "text": data.text ? 
+            i18n.translate("coffeeRoulette.recipientMessageWithText", options ) :
+            i18n.translate("coffeeRoulette.recipientMessage", options),
+        }
+      },
+      {
+        "type": "actions",
+        "elements": [
+          {
+            "type": "button",
+            "style": "primary",
+            "text": {
+              "type": "plain_text",
+              "text": i18n.translate("button.accept"),
+            },
+            "value": data.userId,
+            "action_id": "accept-coffee"
+          },
+          {
+            "type": "button",
+            "style": "danger",
+            "text": {
+              "type": "plain_text",
+              "text": i18n.translate("button.reject"),
+            },
+            "value": data.userId,
+            "action_id": "reject-coffee"
+          }
+        ]
+      }
+    ]
+  
+    return new SlackInteractiveBlock(blocks)
+  }
+
+  static tryAgainCoffeeMessage = async (): Promise<SlackInteractiveBlock> => {
+    const i18n = await I18n.getInstance()
+    
+    const blocks = [
+      {
+        "type": "section",
+        "text": {
+          "type": "mrkdwn",
+          "text": i18n.translate("coffeeRoulette.rejectedOffer")
+        }
+      },
+      {
+        "type": "actions",
+        "elements": [
+          {
+            "type": "button",
+            "style": "primary",
+            "text": {
+              "type": "plain_text",
+              "text": i18n.translate("button.yes")
+            },
+            "action_id": "try-again-coffee"
+          },
+          {
+            "type": "button",
+            "style": "danger",
+            "text": {
+              "type": "plain_text",
+              "text": i18n.translate("button.no")
+            },
+            "action_id": "stop-coffee"
+          }
+        ]
+      }
+    ]
+  
+    return new SlackInteractiveBlock(blocks)
+  }
 
 }
 export class SlackModal extends InteractiveView implements SlackBlock {
